+++ packages/frontend/@n8n/rest-api-client/src/api/index.ts
@@ -1,4 +1,5 @@
 export * from './api-keys';
+export * from './cloudPlans';
 export * from './communityNodes';
 export * from './ctas';
 export * from './eventbus.ee';
